Added custom mediator

+++ src/DanielBlog.Infrastructure/Persistence/Repositories/BlogRepository.cs
@@ -1,4 +1,5 @@
 using DanielBlog.Domain.blogs;
+using DanielBlog.Domain.blogs.Exceptions.BlogExceptions;
 using DanielBlog.Domain.blogs.Interfaces;
 using DanielBlog.Infrastructure.Persistence.DatabaseContext;
 
@@ -6,9 +7,16 @@ namespace DanielBlog.Infrastructure.Persistence.Repositories;
 
 public sealed class BlogRepository(AppDbContext context) : IBlogRepository
 {
-    public async Task<Blog?> GetBlogByIdAsync(Guid id, CancellationToken cancellationToken = default)
+    public async Task<Blog> GetBlogByIdAsync(Guid id, CancellationToken cancellationToken = default)
     {
-        return await context.Blogs.FindAsync(id, cancellationToken);
+        var blog = await context.Blogs.FindAsync(id, cancellationToken);
+        
+        if (blog is null)
+        {
+            throw new BlogNotFoundException($"Blog with id {id} not found.");
+        }
+        
+        return blog;
     }
 
     public async Task<Blog> CreateBlogAsync(Blog blog, CancellationToken cancellationToken)
@@ -22,6 +30,12 @@ public sealed class BlogRepository(AppDbContext context) : IBlogRepository
     public async Task DeleteBlogAsync(Guid id, CancellationToken cancellationToken)
     {
         var blog = await context.Blogs.FindAsync(id, cancellationToken);
+        
+        if (blog is null)
+        {
+            throw new BlogNotFoundException($"Blog with id {id} not found.");
+        }
+        
         context.Blogs.Remove(blog);
         await context.SaveChangesAsync(cancellationToken);
     }
